% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.dataCleaning}
\alias{kasa.dataCleaning}
\title{The checking the data clean}
\usage{
kasa.dataCleaning(x)
}
\arguments{
\item{x}{input dataframe}
}
\value{
results from checking
}
\description{
The checking the data clean
}
