% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.duplicationRemovalBySD}
\alias{kasa.duplicationRemovalBySD}
\title{Duplicated value removal by SD}
\usage{
kasa.duplicationRemovalBySD(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
removed dataframe
}
\description{
Duplicated value removal by SD
}
