% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.geneMedianCentering}
\alias{kasa.geneMedianCentering}
\title{Median centering}
\usage{
kasa.geneMedianCentering(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
Matrix as median centered
}
\description{
Median centering
}
