% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.geneRobustZscoring}
\alias{kasa.geneRobustZscoring}
\title{gene Robust_modified z-scoring}
\usage{
kasa.geneRobustZscoring(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
z-scored genes matrix
}
\description{
gene Robust_modified z-scoring
}
