% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.geneStandardization}
\alias{kasa.geneStandardization}
\title{gene Standardization}
\usage{
kasa.geneStandardization(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
standarized genes
}
\description{
gene Standardization
}
