% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.geneZscoring}
\alias{kasa.geneZscoring}
\title{gene z-scoring}
\usage{
kasa.geneZscoring(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
z-scored genes matrix
}
\description{
gene z-scoring
}
