% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installFunctions.R
\name{kasa.instPak}
\alias{kasa.instPak}
\title{Downloading and intalling the packages automatically from the crans}
\usage{
kasa.instPak(pkg)
}
\arguments{
\item{pkg}{input the vector of the package to be installed #'}
}
\description{
Downloading and intalling the packages automatically from the crans
}
