% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installFunctions.R
\name{kasa.instPak_bioc}
\alias{kasa.instPak_bioc}
\title{Downloading and intalling the packages automatically from the Bioclite}
\usage{
kasa.instPak_bioc(packages_bioc)
}
\arguments{
\item{packages_bioc}{input the vector of the package to be installed #'}
}
\description{
Downloading and intalling the packages automatically from the Bioclite
}
