% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionForDealDatasets.R
\name{kasa.matchingRow}
\alias{kasa.matchingRow}
\title{Two dataset merging and devide into each sets}
\usage{
kasa.matchingRow(dataframeX = x, dataframeY = y, keycolX = "sample",
  keycolY = "sample")
}
\arguments{
\item{dataframeX}{input the dataset X}

\item{dataframeY}{input the dataset Y}

\item{keycolX}{Key column in X}

\item{keycolY}{Key coulum in Y}
}
\value{
List file contains each sets
}
\description{
Two dataset merging and devide into each sets
}
