% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.transform_na_to_median}
\alias{kasa.transform_na_to_median}
\title{Transfomr_NA_to_Median}
\usage{
kasa.transform_na_to_median(x)
}
\arguments{
\item{x}{input dataframe of gene matrix}
}
\value{
Matrix with transformed NA values
}
\description{
Transfomr_NA_to_Median
}
