% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsForGenes.R
\name{kasa.transposeMatrix}
\alias{kasa.transposeMatrix}
\title{Transpose matrix}
\usage{
kasa.transposeMatrix(x, firstColumnName = "sample")
}
\arguments{
\item{x}{input dataframe of gene matrix}

\item{firstColumnName}{input String for the first column name}
}
\value{
transposed matrix as dataframe
}
\description{
Transpose matrix
}
